Added base form for playfair forms
