+++ Cryptographer/frmPlayfairBase.resx
@@ -118,12 +118,9 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <metadata name="openFileDialog1.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
-    <value>17, 17</value>
+    <value>293, 17</value>
   </metadata>
   <metadata name="saveFileDialog1.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
-    <value>157, 17</value>
+    <value>433, 17</value>
   </metadata>
-  <data name="lblDescription2.Text" xml:space="preserve">
-    <value>After that we need to parse the message to make sure the text contains letters from the english alphabet. If everything is ok within the message we pad the key to be the same length as our message. This is done by repeating the keys text until it has the same character length as the message.</value>
-  </data>
 </root>
